refactor: dynamodb external system api
