+++ core/target-systems/dynamodb-external-system-api/src/main/java/io/flamingock/externalsystem/dynamodb/api/DynamoDBExternalSystem.java
@@ -1,4 +1,4 @@
-package io.flamingock.targetsystem.dynamodb;
+package io.flamingock.externalsystem.dynamodb.api;
 
 import io.flamingock.api.external.ExternalSystem;
 import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
